[refactor] multipleCreateOrUpdatePlans mutation 리팩토링
#102
