+++ src/contexts/RecoilProvider/hooks/useSBDOneRM.ts
@@ -1,19 +1,13 @@
-import { useQuery } from '@apollo/client';
 import { useEffect } from 'react';
-import { useSetRecoilState } from 'recoil';
 
-import { GET_SBD_ONE_RM, SBDOneRM } from '@src/operations/queries/getOneRM';
-import { SBDOneRMState } from '@src/recoils';
+import { useSBDOneRM } from '@src/operations/queries/getOneRM';
 
 const useSBDOneRm = (): { loading: boolean } => {
-  const setSBDOneRM = useSetRecoilState<SBDOneRM>(SBDOneRMState);
-  const { data, loading } = useQuery<SBDOneRM>(GET_SBD_ONE_RM);
+  const [getSBDOneRM, { loading }] = useSBDOneRM();
 
   useEffect(() => {
-    if (data) {
-      setSBDOneRM(data);
-    }
-  }, [data, setSBDOneRM]);
+    getSBDOneRM();
+  }, [getSBDOneRM]);
 
   return { loading };
 };
