+++ src/operations/mutations/multipleCreateOrUpdatePlans.ts
@@ -1,12 +1,16 @@
 import { gql, useMutation } from '@apollo/client';
 import { FetchResult } from '@apollo/client/link/core';
 import { MutationFunctionOptions, MutationResult } from '@apollo/client/react';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 
 import { CORE_PLAN_FIELDS } from '@src/fragments/plan';
-import useErrorEffect from '@src/hooks/useErrorEffect';
+import { useSBDOneRM } from '@src/operations/queries/getOneRM';
+import { plansState } from '@src/recoils';
+import { deletePlansState } from '@src/screens/EditPlanScreen/recoils';
 import {
   Mutation,
   MutationMultipleCreateOrUpdatePlansArgs,
+  Plan,
 } from '@src/types/graphql';
 
 export const MULTIPLE_CREATE_OR_UPDATE_PLANS = gql`
@@ -18,7 +22,9 @@ export const MULTIPLE_CREATE_OR_UPDATE_PLANS = gql`
   }
 `;
 
-export const useMultipleCreateOrUpdatePlansMutation = (): [
+export const useMultipleCreateOrUpdatePlansMutation = (
+  callback: () => any | undefined = () => null,
+): [
   (
     options?: MutationFunctionOptions<
       Pick<Mutation, 'multipleCreateOrUpdatePlans'>,
@@ -27,15 +33,30 @@ export const useMultipleCreateOrUpdatePlansMutation = (): [
   ) => Promise<FetchResult<Pick<Mutation, 'multipleCreateOrUpdatePlans'>>>,
   Pick<
     MutationResult<Pick<Mutation, 'multipleCreateOrUpdatePlans'>>,
-    'data' | 'loading'
+    'loading'
   >,
 ] => {
-  const [multipleCreateOrUpdatePlans, { data, error, loading }] = useMutation<
+  const setPlans = useSetRecoilState<Plan[]>(plansState);
+  const deletePlans = useRecoilValue<string[]>(deletePlansState);
+  const [getSBDOneRM] = useSBDOneRM(callback);
+  const [multipleCreateOrUpdatePlans, { loading }] = useMutation<
     Pick<Mutation, 'multipleCreateOrUpdatePlans'>,
     MutationMultipleCreateOrUpdatePlansArgs
-  >(MULTIPLE_CREATE_OR_UPDATE_PLANS);
+  >(MULTIPLE_CREATE_OR_UPDATE_PLANS, {
+    onCompleted: data => {
+      setPlans(prevPlans => {
+        const plansIds = (data.multipleCreateOrUpdatePlans as Plan[]).map(
+          ({ _id }) => _id,
+        );
 
-  useErrorEffect(error);
+        return prevPlans
+          .filter(({ _id }) => !deletePlans.includes(_id))
+          .filter(({ _id }) => !plansIds.includes(_id))
+          .concat(data.multipleCreateOrUpdatePlans);
+      });
+      getSBDOneRM();
+    },
+  });
 
-  return [multipleCreateOrUpdatePlans, { data, loading }];
+  return [multipleCreateOrUpdatePlans, { loading }];
 };
